requirements.txt
setproctitle==1.1.10
PyYAML==3.13
influxdb==5.2.1